# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "业务审查任务开放 API",
    "description": "统一任务接口，支持 **classification**（材料分类）、**extraction**（要素抽取）、**due_diligence**（债权尽调填报）、**template_fill**（模板驱动填报）。对外集成请使用 `/api/v1/*` 并携带 API Key。",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "v1-tasks"
        ],
        "summary": "任务列表",
        "description": "按创建时间倒序返回任务；可用 `task_type` 过滤。任务类型：`classification`（材料分类）、`extraction`（要素抽取）、`due_diligence`（债权尽调指标填报）、`template_fill`（模板驱动填报）",
        "operationId": "v1_list_tasks_api_v1_tasks_get",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "task_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Type"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskListItemV1"
                  },
                  "title": "Response V1 List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "服务端未配置 API_KEYS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "v1-tasks"
        ],
        "summary": "统一创建任务",
        "description": "通过 `task_type` 选择任务类型并上传对应材料。任务类型：`classification`（材料分类）、`extraction`（要素抽取）、`due_diligence`（债权尽调指标填报）、`template_fill`（模板驱动填报） `classification` 需 `files`、`classification_basis`；`extraction` 需 `files`、`extract_schema`；`due_diligence` 需 `due_diligence_docx`；`template_fill` 需 `source_files` 与 `collection_template`。",
        "operationId": "v1_create_task_api_v1_tasks_post",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v1_create_task_api_v1_tasks_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "服务端未配置 API_KEYS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "task_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "general",
                    "classification",
                    "extraction",
                    "case1",
                    "case2"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskListItem"
                  },
                  "title": "Response List Tasks Api Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_api_tasks_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/case1/tasks": {
      "post": {
        "tags": [
          "case1"
        ],
        "summary": "Create Case1 Task Endpoint",
        "operationId": "create_case1_task_endpoint_api_case1_tasks_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_case1_task_endpoint_api_case1_tasks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/case2/tasks": {
      "post": {
        "tags": [
          "case2"
        ],
        "summary": "Create Case2 Task Endpoint",
        "operationId": "create_case2_task_endpoint_api_case2_tasks_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_case2_task_endpoint_api_case2_tasks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_case1_task_endpoint_api_case1_tasks_post": {
        "properties": {
          "due_diligence_docx": {
            "type": "string",
            "format": "binary",
            "title": "Due Diligence Docx"
          },
          "supplementary_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplementary Files"
          },
          "run_ocr": {
            "type": "string",
            "title": "Run Ocr",
            "default": "false"
          },
          "indicator_judgment_rules": {
            "type": "string",
            "title": "Indicator Judgment Rules",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "due_diligence_docx"
        ],
        "title": "Body_create_case1_task_endpoint_api_case1_tasks_post"
      },
      "Body_create_case2_task_endpoint_api_case2_tasks_post": {
        "properties": {
          "source_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Files"
          },
          "collection_template": {
            "type": "string",
            "format": "binary",
            "title": "Collection Template"
          },
          "run_ocr": {
            "type": "string",
            "title": "Run Ocr",
            "default": "true"
          },
          "fill_logic_rules": {
            "type": "string",
            "title": "Fill Logic Rules",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "collection_template"
        ],
        "title": "Body_create_case2_task_endpoint_api_case2_tasks_post"
      },
      "Body_create_task_api_tasks_post": {
        "properties": {
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "classification_basis": {
            "type": "string",
            "title": "Classification Basis"
          },
          "extract_schema": {
            "type": "string",
            "title": "Extract Schema"
          },
          "collection_template": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Template"
          },
          "run_ocr": {
            "type": "string",
            "title": "Run Ocr",
            "default": "true"
          },
          "run_agent": {
            "type": "string",
            "title": "Run Agent",
            "default": "true"
          },
          "run_collection_fill": {
            "type": "string",
            "title": "Run Collection Fill",
            "default": "true"
          }
        },
        "type": "object",
        "required": [
          "classification_basis",
          "extract_schema"
        ],
        "title": "Body_create_task_api_tasks_post"
      },
      "Body_v1_create_task_api_v1_tasks_post": {
        "properties": {
          "task_type": {
            "$ref": "#/components/schemas/TaskType",
            "description": "任务类型：`classification`（材料分类）、`extraction`（要素抽取）、`due_diligence`（债权尽调指标填报）、`template_fill`（模板驱动填报）"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files",
            "description": "classification / extraction：材料文件（可多选）"
          },
          "classification_basis": {
            "type": "string",
            "title": "Classification Basis",
            "description": "classification：分类依据",
            "default": ""
          },
          "extract_schema": {
            "type": "string",
            "title": "Extract Schema",
            "description": "extraction：抽取 schema（JSON 字符串）",
            "default": ""
          },
          "collection_template": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Template",
            "description": "template_fill：填报模板 xlsx"
          },
          "run_ocr": {
            "type": "string",
            "title": "Run Ocr",
            "default": "true"
          },
          "due_diligence_docx": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Diligence Docx",
            "description": "due_diligence：尽调 docx"
          },
          "supplementary_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplementary Files",
            "description": "due_diligence：补充材料"
          },
          "indicator_judgment_rules": {
            "type": "string",
            "title": "Indicator Judgment Rules",
            "description": "due_diligence：指标判断规则",
            "default": ""
          },
          "source_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Files",
            "description": "template_fill：源文件"
          },
          "fill_logic_rules": {
            "type": "string",
            "title": "Fill Logic Rules",
            "description": "template_fill：填表逻辑与计算规则",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "task_type"
        ],
        "title": "Body_v1_create_task_api_v1_tasks_post"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TaskCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "TaskCreateResponse"
      },
      "TaskListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "zip_filename": {
            "type": "string",
            "title": "Zip Filename"
          },
          "task_kind": {
            "type": "string",
            "title": "Task Kind",
            "default": "general"
          },
          "source_docx_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Docx Filename"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "zip_filename",
          "created_at"
        ],
        "title": "TaskListItem"
      },
      "TaskListItemV1": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "zip_filename": {
            "type": "string",
            "title": "Zip Filename"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "source_docx_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Docx Filename"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "zip_filename",
          "task_type",
          "created_at"
        ],
        "title": "TaskListItemV1"
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "classification",
          "extraction",
          "due_diligence",
          "template_fill"
        ],
        "title": "TaskType",
        "description": "对外 API `/api/v1` 的 task_type 取值。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "在 `.env` 中配置 `API_KEYS`；亦支持 `Authorization: Bearer <key>`"
      }
    }
  },
...
}